started ex03_02, minor changes to ex03_01
Hey guys,

I started with the second part of the homework, but I could not finish
it yet. I find it really hard to understand what exactly they are doing
in the paper, so I searched for a example implementation and found this:

http://www.mathworks.com/matlabcentral/fileexchange/17894-keypoint-extraction/content/keypointExtraction%5Ckp_harrislaplace.m

In my opinion, this example makes it much more clear what should be
happening.
As I said, I just began working through the implementation but can't
proceed for now, so I wanted to commit my changes so far.
Also I did some really minor changes to the first exercise, nothing
special ^^. If I overlooked something and the changes are wrong, just
tell me.

Best regards,
Daniel