Started warping/backwarping process
I started the the warping/backwarping process for the robust feature
points. At this stage, i produce random affine transformations  with the
formula in the paper. The formula is written in line 31 of the code, the
distribution of the angles/lambdas is taken from the paper, too.
Then the patches around each harrisPoint are warped a number of times
(backwarps times) and thus new patches are created. The regions of the
patches where no warped pixel was projected at are then filled with
random black&white noise, as seen in the lecture slides.
I don't know if the code works like this and is correct and I'm not even
sure if that's the right approach.
At next I will look to finding new harrisPoints in the warped image
patches and then backproject these points to find out if they are robust
enough.
I guess for the warping process for FERN training we can use most of the
codes I am writing here.